% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.cluster.dist}
\alias{check.cluster.dist}
\title{Perform checks that cluster and dist are compatible with each other}
\usage{
check.cluster.dist(cluster, dist)
}
\arguments{
\item{cluster}{vector with cluster assignments, must be named}

\item{dist}{dist or matrix, must match cluster}
}
\value{
matrix, a subset of dist
}
\description{
(This is an internal function only)
}
