% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check.numeric}
\alias{check.numeric}
\title{performs checks on an input to make sure it is in the required range}
\usage{
check.numeric(x, range = c(0, 1), varname = "threshold")
}
\arguments{
\item{x}{numeric value}

\item{range}{pair of minimum, maximum values}

\item{varname}{character, printed during stop()}
}
\value{
numeric x
}
\description{
(This is an internal function only)
}
