% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medinfo.R
\name{dist2centers}
\alias{dist2centers}
\title{compute a matrix of distances from data elements to fixed points
(e.g. cluster centers)}
\usage{
dist2centers(data, centers)
}
\arguments{
\item{data}{matrix with raw data}

\item{centers}{matrix with fewer rows than data each row meant to
capture center of a cluster}
}
\value{
matrix
}
\description{
(This is an internal function only)
}
