% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdata.R
\docType{data}
\name{kdata.1}
\alias{kdata.1}
\title{Toy dataset with two convex groups with partial overlap}
\format{matrix with two columns: \code{D1}, \code{D2}}
\usage{
data(kdata.1)
}
\description{
Toy dataset with two convex groups with partial overlap
}
\keyword{datasets}
