% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdata.R
\docType{data}
\name{kdata.2}
\alias{kdata.2}
\title{Toy dataset with two non-overalpping and non-spherical groups}
\format{matrix with two columns: \code{D1}, \code{D2}}
\usage{
data(kdata.2)
}
\description{
Toy dataset with two non-overalpping and non-spherical groups
}
\keyword{datasets}
