% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdata.R
\docType{data}
\name{kdata.3}
\alias{kdata.3}
\title{Toy dataset with three groups}
\format{matrix with two columns: \code{D1}, \code{D2}}
\usage{
data(kdata.3)
}
\description{
Toy dataset with three groups
}
\keyword{datasets}
