% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdata.R
\docType{data}
\name{kdata.4}
\alias{kdata.4}
\title{Toy dataset with four groups atop a wide area of noise points}
\format{matrix with two columns: \code{D1}, \code{D2}}
\usage{
data(kdata.4)
}
\description{
Toy dataset with four groups atop a wide area of noise points
}
\keyword{datasets}
