% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksharp.R
\name{ksharp.prep}
\alias{ksharp.prep}
\title{augment a kmeans object with sharpness values}
\usage{
ksharp.prep(x, data, cluster)
}
\arguments{
\item{x}{object of class kmeans or cluster}

\item{data}{matrix with raw data, must match x}

\item{cluster}{named vector assigning items in data to clusters;
these clusters are meant to be the raw/original/unsharpened clusters}
}
\value{
object based on x, with additional elements
}
\description{
(This is an internal function)
}
