% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksharp.R
\name{ksharp}
\alias{ksharp}
\title{sharpen a clustering}
\usage{
ksharp(
  x,
  threshold = 0.1,
  data = NULL,
  method = c("silhouette", "neighbor", "medoid"),
  threshold.abs = NULL
)
}
\arguments{
\item{x}{clustering object; several types of inputs are acceptable, including
objects of class kmeans, pam, and self-made lists with a component "cluster".}

\item{threshold}{numeric; the fraction of points to place in noise group}

\item{data}{matrix, raw data corresponding to clustering x; must be present
when sharpening for the first time or if data is not present within x.}

\item{method}{character, determines method used for sharpening}

\item{threshold.abs}{numeric; absolute-value of threshold for sharpening.
When non-NULL, this value overrides value in argument 'threshold'}
}
\description{
Each data point in a clustering is assigned to a cluster, but some data points
may lie in ambiguous zones between two or more clusters. Cluster sharpening
assigns these border points into a separate noise group, thereby creating
more stark distinctions between the groupings that are left.
}
\details{
Noise points are assigned to a group with cluster index 0. This is analogous
behavior to output produced by dbscan.
}
