% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medinfo.R
\name{medinfo}
\alias{medinfo}
\title{compute info on distances to medoids/centroids}
\usage{
medinfo(cluster, data, silwidths)
}
\arguments{
\item{cluster}{named vector}

\item{data}{matrix}

\item{silwidths}{matrix with silhouette widths}
}
\value{
list with component widths
}
\description{
Analogous in structure to silinfo and neiinfo, it computes
a "widths" marix assessing how well each data point belongs
to its cluster. This measure is here the ratio of the distance
from the point to the nearest cluster center to the distance
from the point to its own cluster center.
}
