% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medinfo.R
\name{medoids}
\alias{medoids}
\title{get coordinates of cluster medoids}
\usage{
medoids(data, silwidths)
}
\arguments{
\item{data}{matrix}

\item{silwidths}{matrix with silhouette widths}
}
\value{
matrix, each row has an element of data that represents its cluster
}
\description{
(This is an internal functions)
}
